Move folders to modules and update imports
